seting initial x, y to with/2 and height/2 was casting x and y as floats, changed to int(width/2) and int(height/2)

+++ tk_test.py
@@ -6,8 +6,8 @@ FG_COLOR = "red"
 
 world_width = 600
 world_height = 500
-x = world_width / 2
-y = world_height / 2
+x = int(world_width / 2)
+y = int(world_height / 2)
 xmove = 10
 ymove = 0
 last_xmove = None
